requirements.txt
numpy==1.19.1
pandas==1.1.3
psycopg2==2.8.5
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
xlrd==1.2.0
